template within template, sidebar, sample content
- Template within template:
we now have a drug.html template for each drug, which has base.html as
it’s template. drug.html is a template for alcohol.html

- sidebar:
We have a scrolling sidebar like the one on the bootstrap documentation
site.

- sample content:
I plugged in some lorem ipsum to try out the template-ception and
scrolling sidebar.